chore(event-handler): split Router tests into multiple files
